Show correct server in switch notifications [VPNAND-883].
Don't use Profile.server in buildReconnectionInfo because it may use a
stale value of VpnUser to select the server (e.g. when it's right after
a downgrade).
Instead the server is chosen by the logic in VpnConnectionErrorHandler
and transferred with the VpnFallbackResult.Switch object. This way the
same value is used to show notification and to connect.