Accept location response only if VPN is off [VPNAND-659].
The request to get client IP is only started when VPN is off, otherwise
the returned IP will be one of the VPN servers.
It's possible, at least in theory, that VPN connection is started right
after making the request. With the retry mechanism it could happen that
a successful request is made through VPN returning a wrong IP.

To work around this the result is only accepted if the VPN is still
disabled.