Introduce Profile comparison [VPNAND-950].
Checking which Profile is connected to (e.g. in Profiles tab) is now
done by comparing the Profiles and not Profile.server.

Additionally, isSecureCore flag is moved from ServerWrapper to Profile
and it may now be "null" indicating the Profile works with Secure Core
and regular servers depending on the global setting.