Store country codes in TV recents [VPNAND-958].
The RecentsManager uses Profile objects to store the list of countries
recently connected to. Unfortunately the Profiles point to individual
servers, so connecting to a recent country connects to previous servers,
not the best server for that country.
With this change the RecentsManager stores a list of country codes which
fixes the issue.