Avoid fallback logic when switching Secure Core [VPNAND-1005].
When switching Secure Core while connected, the app reconnects to a
server in the same country. Not every country is available in both
modes. If this happens, the attempt to connect will fail and trigger
recovery logic that in the end shows the user a notification.

To avoid this getReconnectProfileOnSecureCoreChange checks if a sever
for the new profile exists and uses fallback profile if it doesn't.