Prevent loopback IPs from appearing in allowed IPs [VPNAND-930].
Adding any route starting with 127.* to VpnService.Builder.addRoute()
causes an exception to be thrown. It's possible to for
calculateAllowedIps to generate a range with 127 if excluded IPs are
adjacent to this range.
In this case 127.0.0.0/8 is added to excludes. It is not added for all
calculations to avoid making the allowed IP ranges unnecessarily complex
for 99% of cases.