Fix CI build:
Remove the .ndk base job and revert the before_script to the default
section. Having an .ndk job that most jobs need to extend is going to be
a maintenance nightmare.
Deploy jobs have been added with rules and "needs" dependencies so they
can run in parallel with other jobs (esp. build).